chore(ui): Split up compliance ui services
