+++ ui/apps/platform/src/Components/PatternFly/DayPickerDropdown.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement } from 'react';
 import { Select, SelectOption, SelectVariant } from '@patternfly/react-core/deprecated';
 
 import useMultiSelect from 'hooks/useMultiSelect';
-import { IntervalType } from 'services/ComplianceEnhancedService';
+import { IntervalType } from 'services/ComplianceScanConfigurationService';
 
 export type DayPickerDropdownProps = {
     fieldId: string;
