+++ ui/apps/platform/src/Containers/ComplianceEnhanced/Schedules/Wizard/ProfileSelection.tsx
@@ -17,7 +17,7 @@ import { SearchIcon } from '@patternfly/react-icons';
 
 import EmptyStateTemplate from 'Components/EmptyStateTemplate';
 import useTableSelection from 'hooks/useTableSelection';
-import { ComplianceProfileSummary } from 'services/ComplianceEnhancedService';
+import { ComplianceProfileSummary } from 'services/ComplianceProfileService';
 
 import { ScanConfigFormValues } from '../compliance.scanConfigs.utils';
