+++ ui/apps/platform/src/Containers/ComplianceEnhanced/Schedules/compliance.scanConfigs.utils.test.ts
@@ -1,4 +1,4 @@
-import { Schedule } from 'services/ComplianceEnhancedService';
+import { Schedule } from 'services/ComplianceScanConfigurationService';
 import {
     convertFormikParametersToSchedule,
     convertScheduleToFormikParameters,
