+++ ui/apps/platform/src/Containers/ComplianceEnhanced/Schedules/components/ScanConfigClustersTable.tsx
@@ -4,7 +4,7 @@ import React, { useState } from 'react';
 import { Card, CardBody, CardHeader, CardTitle, Pagination } from '@patternfly/react-core';
 import { Table, Tbody, Td, Th, ThProps, Thead, Tr } from '@patternfly/react-table';
 
-import { ClusterScanStatus } from 'services/ComplianceEnhancedService';
+import { ClusterScanStatus } from 'services/ComplianceScanConfigurationService';
 
 import ComplianceClusterStatus from './ComplianceClusterStatus';
