+++ ui/apps/platform/src/services/ComplianceCommon.ts
@@ -1,0 +1,1 @@
+export const complianceV2Url = '/v2/compliance';
